feature/#207 - unified and simplified the attribute importance settings
Deleted file:
* `user_preferences_view.dart`: now it's a page - `UserPreferencesPage`

New files:
* `attribute_button.dart`: Colored button for attribute importance, with corresponding action
* `user_preferences_page.dart`: Preferences page for attribute importances

Impacted files:
* `home_page.dart`: moved code to new widget `AttributeButton`; replaced `UserPreferencesView` with `UserPreferencesPage`
* `personalized_ranking_page.dart`: replaced `UserPreferencesView` with `UserPreferencesPage`
* `product_page.dart`: replaced `UserPreferencesView` with `UserPreferencesPage`
* `settings_page.dart`: replaced `UserPreferencesView` with `UserPreferencesPage`
* `user_preferences.dart`: removed the notion of collapsed attribute group in preferences
