+++ packages/smooth_app/lib/pages/product/product_page.dart
@@ -18,7 +18,7 @@ import 'package:smooth_app/data_models/product_preferences.dart';
 import 'package:smooth_ui_library/widgets/smooth_card.dart';
 
 // Project imports:
-import 'package:smooth_app/bottom_sheet_views/user_preferences_view.dart';
+import 'package:smooth_app/pages/user_preferences_page.dart';
 import 'package:smooth_app/cards/data_cards/image_upload_card.dart';
 import 'package:smooth_app/cards/expandables/attribute_list_expandable.dart';
 import 'package:smooth_app/data_models/product_list.dart';
@@ -114,7 +114,13 @@ class _ProductPageState extends State<ProductPage> {
           onTap: (final int index) async {
             switch (index) {
               case 0:
-                UserPreferencesView.showModal(context);
+                await Navigator.push<Widget>(
+                  context,
+                  MaterialPageRoute<Widget>(
+                    builder: (BuildContext context) =>
+                        const UserPreferencesPage(),
+                  ),
+                );
                 return;
               case 1:
                 await _copy(
